backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
biopython==1.84
requests==2.32.3
python-dotenv==1.0.1
numpy
